refactor: Rename `lib/plugins/aws/deploy/lib/uploadArtifacts.js`
