+++ lib/plugins/aws/deploy/index.js
@@ -12,7 +12,7 @@ const getExecuteChangeSetParams = require('../lib/get-execute-change-set-params'
 const waitForChangeSetCreation = require('../lib/wait-for-change-set-creation');
 const createStack = require('./lib/create-stack');
 const cleanupS3Bucket = require('./lib/cleanup-s3-bucket');
-const uploadArtifacts = require('./lib/uploadArtifacts');
+const uploadArtifacts = require('./lib/upload-artifacts');
 const validateTemplate = require('./lib/validate-template');
 const updateStack = require('../lib/update-stack');
 const ensureValidBucketExists = require('./lib/ensure-valid-bucket-exists');
