+++ lib/plugins/aws/deploy/lib/upload-artifacts.js
